Improve output of test runner a bit
Try to make output of the test driver a bit more readable. Still need a
way to determine if the tokenizer or the evaluator is expected to fail
in case of tests marked with an error code. The error code might give a
clue: some are only generated by the tokenizer, not the evaluator.